+++ docs/changes.rst
@@ -6,7 +6,10 @@ Version 0.3.6
 
 To be released.
 
-- Fix :exc:`UnicodeEncodeError` when :meth:`Drawing.text()
+- Added :meth:`Drawing.rectangle() <wand.drawing.Drawing.rectangle>` method.
+  [:issue:`159`]
+- Added :func:`contextlib.nested()` function to :meth:`wand.compat` module.
+- Fixed :exc:`UnicodeEncodeError` when :meth:`Drawing.text()
   <wand.drawing.Drawing.text>` method gives Unicode ``text`` argument
   in Python 2.  [:issue:`163`]